有了一个新页面，开始做页面的简单版本，更完就去睡觉了

+++ src/main/java/top/yzlin/teamraise/parse/AlreadyRaise.java
@@ -5,9 +5,9 @@ import top.yzlin.teamraise.entity.RaiseInfo;
 
 public interface AlreadyRaise {
 
-    default int getMoney(RaiseInfo raiseInfo){
+    default double getMoney(RaiseInfo raiseInfo){
         return getMoney(raiseInfo.getProID());
     }
 
-    int getMoney(int proID);
+    double getMoney(int proID);
 }
